Rusty Skylines Alpha v1.0.7a
After a rough week, I started making outlines for the gui elements.
The gui layers have an 'opaque' option now!

If you press on a circle, it draws a dark circle on it as a selection indicator, pretty crude, but it will be a beautifully animated border later!

Had to solve some blending issues for the circle quad being transparent with no blending, but I sorted that out, ask me how.

Anyway, you don't have to worry about this stuff, I do! So just enjoy this engine in disguise as a game!

I like writing commit messages...
I could write 3x more, but do you want that? Who tf is reading this anyway but me? This has 0 Stars and views on GitHub!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>