Rusty Skylines Alpha v1.0.8a
I added a circle outline gui element type, which draws an animated dashed line with sub dashed lines inside, a bit buggy still, but will be fixed hopefully tomorrow!

If you press on a circle, it draws an outline on it as a selection indicator, pretty crude still, but it will be an even more beautifully animated outline!

Anyway, you don't have to worry about this stuff, I do! So just enjoy this engine in disguise as a game!

I like writing commit messages...
I could write 3x more, but do you want that? Who tf is reading this anyway but me? This has 0 Stars and views on GitHub!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>