Rusty Skylines Alpha v1.0.9a
Simply made the circle outline shader look good ENOUGH! Not perfect, but alright and pretty customizable!

Fixed some alignment issues with the data passed on to the shader, which caused the sub-dash color to be wrong, that's because it took the data from the 'misc' vec4 instead of the sub-dash color!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>