Rusty Skylines Alpha v1.0.10a
Cat is sitting on my desk.

Other than that, I made deselecting possible!

I added a handle element which spawns if you select a circle (only circle for now) and follows the mouse as best as possible. There is a main handle and sub handle for drawing, it has no functionality for now!
It fades into thin air if you move away with your mouse...

Signed-off-by: maxwag9 <maximwagner46@gmail.com>