Rusty Skylines Alpha v1.1.0a
The handle can be moved!!
And it feels good to move, with hover and click detection and smooth resizing!

Also, I fixed some issues with color blending (alpha) in the ui_handle.wgsl shader.

Oh and I forgot, I updated the circle outline shader to resize all its parameters based on its radius, so the thickness and length of the dashes remain constant. I must say those sub dashes walking around do look cool!

My rustrover trial is ending in 5 days, but that is cap, because rustrover is free for non-commercial users, and I am not selling yet!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>