Rusty Skylines Alpha v1.1.1a
Moving polygons and the holy square is now possible in editor mode, that includes selecting the polygons!

The selection outline of the polygons looks very buggy, but I will fix it soon.

I removed all mentions of rectangles and triangles, because they are just glorified polygons...

I am going crazy.

For convenience, I added a keybind to toggle editor_mode with F6!

My rustrover trial is ending in 5days, but that is cap!, because rustrover is free for non-commercial users, and I am not selling this gui editor 'game' yet!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>