Rusty Skylines Alpha v1.1.2a
The polygon outline looks a little less buggy.

AND you can move single vertices now!!
I am going crazy.

For convenience, I added a keybind to toggle editor_mode with F6!

My rustrover trial is ending in 1 day..., but that is cap!!, because rustrover is free for non-commercial users, and I am not selling this gui editor 'game' yet! valvE

Signed-off-by: maxwag9 <maximwagner46@gmail.com>