Rusty Skylines Alpha v1.1.3b
Added Z-Ordering! Was simpler than I thought...
That just means that you can choose if a gui element is above or below another gui element.

ALSO, I changed the Background color to (r: 0.0, g: 0.0, b: 0.3, a: 1.0) which should give us a break from the same dull color every day!

Minor, but cool:
1. Circle glow doesn't sit at 0.6 alpha permanently, even if off anymore.
2. Added a fade to the circles, if the fade.x is over 0.5. The circle fades away as the mouse goes away, and it looks like lighting, you can't achieve that on the cpu side!

Now, it all works pretty good and better, but I am (still) not too happy with how vertices are sorted while I am dragging the vertices through the roof and into hell and heaven! Try it for yourself... thx
(I haven't changed the sorting code-1)

Actually, the sorter is kind of in the "Off" state, because it is just a function which reverses the vertices if they are somehow wrong, but ccw sorting doesn't do it better...

For convenience, way long ago I added a keybind to toggle editor_mode with F6!

My rustrover trial is ending in 1 day..., but that is cap!! (Still 1 day? idk man it is 02:03 how is it still an active trial?! UPDATE b update: still 2 day and it is read in the bottom right, idk, doomsday will happen I guess!), because rustrover is free for non-commercial users, and I am not selling this gui editor 'game' yet! valvE

Signed-off-by: maxwag9 <maximwagner46@gmail.com>