Rusty Skylines Alpha v1.1.4b
Now you can SAVE your hard work on the GUI! Just press Strg and S (Or Ctrl S for non-german people) just like in all other programs to save the gui AS-IS, WYSIWYG!

I needed to block the wasdqe movement of the 3D camera to prevent it from moving with "S", but this is only in editor mode anyway, so it of course still works in normal gameplay (By gameplay I mean 'Orbiting a 3D quad and moving around whilst staring at the GUI' simulator!)

Important side-note: I added a way to select only the top most element, instead of selecting all the elements below it as well, but that only works between the same types of elements and not for circles.

For convenience, way, way long ago I added a keybind to toggle editor_mode with F6! This is the last time I am saying this, so remember this!!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>