Rusty Skylines Alpha v1.1.5a
Quick commit: Added just the infancy of a keybinds system, going to work more on it soon.

Also before that added a way to spawn circles, but now it only spawns polygons, because I said so.
For that I also added defaults for every GUI element type!

Sorry for not doing many commits.
I will COMMIT to doing more COMMITS and more PROGRESS!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>