Rusty Skylines Alpha v1.1.6a
Moved all shaders to the all-new data folder.
Made a shader watcher that watches the shader files .wgsl very intently until one changes and the entire pipeline gets rebuilt with the new shader!

Yes, that means you can edit shader on the fly without having to restart the game. It even shows it in-game as text that a shader has been changed!

Signed-off-by: maxwag9 <maximwagner46@gmail.com>