Rusty Skylines Alpha v1.2.7a Dec 5nd 22:41
I seriously have to stop making LAST MINUTE commits in a half-broken state, because it just gives me stress and I fix the last issue just 12 minutes after!

Fixed: When clicking on the HUe Circle, it saved its color to the variables, but that colorful circle is not there to be selected, its color is just gray!
So that's why you got that weird unsatisfying behaviour of the handle returning to the center for no reason!

Worked on the input_box further.
Added two new variables: override_mode and show_gui

Fixed the outline of circles when their border thickness is 0.

Signed-off-by: maxwag9 <maximwagner46@gmail.com>