Rusty Skylines Alpha v1.3.6a Dec 16th 12:13
Fixed frustum culling close to the camera, now chunks don't just vanish right in front of you.

The rendering of the chunks is now supposed to be way more performant because of the new rendering system.
It is now batching chunks into  multiple buffers, so the gpu doesn't need to switch buffers every chunk and can basically just stream the chunks!

Fixed some input box things, like the value not being shown live.

***
He sits to the left of me in a dark green Calvin Klein shirt playing with the cat, and occasionally watching me code in awe, having never written a single line of code in his life.

Signed-off-by: maxwag9 <maximwagner46@gmail.com>