Rusty Skylines Alpha v1.3.6a Dec 16th 16:19 Now ray casting... I added ray casting from the mouse to the terrain and it shows a red circle kind of, just colored the vertices of the terrain red.
Added restrictions to the camera movement: The camera target (origin gizmo) now always stays above ground. The camera pitch is brought up if it is intersecting with any terrain. SO now flying feels a bit smoother.

Made the camera speed with shift way slower, but overall faster because of zooming out speed increase, zooming in is more restricted now, and the speed remains at least 2.0 at full zoom-in.

The ray casting is making the game laggy if the point is far away or the sky... I have to fix this soon!

Cat sleeps. Now cat licks.

Signed-off-by: maxwag9 <maximwagner46@gmail.com>