Rusty Skylines Alpha v1.3.8a_test Dec 20th 20:08
Spent hours making a shitty solution for the z-fighting between water and terrain. It doesn't fix the issue and overcomplicates things by using a depth texture, which doesn't have enough precision, so it inevitable that:
I will switch to depth stencils instead in the next update, which should be PERFECT without ANY wobbling or z-fighting!

Excited to see that!

Halved the render and generation distance.

***
Depth stencils blew my mind when I heard of them, finally a way to fix this shit. This has been in my game for weeks >:)

Signed-off-by: maxwag9 <maximwagner46@gmail.com>