Rusty Skylines Alpha v1.3.11b Dec 23nd 17:16
Improved the performance of terrain editing drastically, though the terrain isn't saved yet and still buggy.
It is very much still buggy and in beta 2.0

Improved the performance of the terrain generation functions drastically, by like 2x-4x!
Because of that I doubled the batch sizes of terrain generation.

Added some debug timing prints for the terrain editing, just temporary.

***
RTX 4060 is mine.

Signed-off-by: maxwag9 <maximwagner46@gmail.com>